table-browser: move table origin info into separate component
