+++ src/scripts/modules/table-browser/react/components/GeneralInfoTab.jsx
@@ -7,12 +7,7 @@ import {Table} from 'react-bootstrap';
 import immutableMixin from '../../../../react/mixins/ImmutableRendererMixin';
 import EmptyState from '../../../components/react/components/ComponentEmptyState';
 import filesize from 'filesize';
-import ComponentsStore from '../../../components/stores/ComponentsStore';
-import ComponentIcon from '../../../../react/common/ComponentIcon';
-import ComponentConfigurationLink from '../../../components/react/components/ComponentConfigurationLink';
-import InstalledComponentsStore from '../../../components/stores/InstalledComponentsStore';
-import {fromJS} from 'immutable';
-
+import TableOriginInfo from '../../../../react/common/TableOriginInfo';
 
 export default React.createClass({
 
@@ -58,31 +53,13 @@ export default React.createClass({
             {this.renderTableRow('Primary Key', _.isEmpty(primaryKey) ? 'N/A' : primaryKey.join(', '))}
             {this.renderTableRow('Last Import', this.renderTimefromNow(table.get('lastImportDate')))}
             {this.renderTableRow('Last Change', this.renderTimefromNow(table.get('lastChangeDate')))}
-            {this.renderTableRow('Origin', this.renderComponentFromMetadata(table))}
+            {this.renderTableRow('Origin', <TableOriginInfo table={table}/>)}
             {this.renderTableRow('Rows Count', this.renderRowsCount(table.get('rowsCount')))}
             {this.renderTableRow('Data Size', this.renderDataSize(table.get('dataSizeBytes')))}
             {this.renderTableRow('Columns', table.get('columns').count() + ' columns: ' + table.get('columns').join(', '))}
           </tbody>
         </Table>
       </div>
-    );
-  },
-
-  renderComponentFromMetadata(table) {
-    const metadata = table.get('metadata');
-    const componentId = metadata.find(m => m.get('key') === 'KBC.lastUpdatedBy.component.id').get('value');
-    const configId = metadata.find(m => m.get('key') === 'KBC.lastUpdatedBy.configuration.id').get('value');
-    const component = componentId && ComponentsStore.getComponent(componentId);
-    if (!component) return 'N/A';
-    const componentName = component.get('type') !== 'transformation' ? `${component.get('name')} ${component.get('type')}` : `${component.get('type')}`;
-    const config = InstalledComponentsStore.getConfig(componentId, configId);
-    const configName = config ? config.get('name', configId) : configId;
-    return (
-      <span>
-        <ComponentIcon component={fromJS(component)}/>
-        <ComponentConfigurationLink componentId={componentId} configId={configId}>{componentName} / {configName}
-        </ComponentConfigurationLink>
-      </span>
     );
   },
